backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==11.1.0
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.31.0
tweepy>=4.15.0
